Decouple the code to make re-use easier.
GrobidProperties is global, and is called in many different places.  This change attempts to isolate dependencies for the various parsers, such that callers can supply the dependencies explicitly.  There is still more work to do for models, but this makes it easier to specify the location of pdftoxml(.exe), and the temp path.

HeaderParser, Segmentation, BookStructureParser - reduce the duplication in this code by moving out the code necessary to convert a PDF to XML, parse the XML, and convert the XML to tokens.  Still more work to do.

EngineParsers - make it possible to inject dependencies for converting PDF files to a Grobid Document type.  Still more work to do.

PdfToXmlConverter - interface to centralize exactly how you want to extract content from a document.  I _think_ there are three different flavors.
   1. Extract headers, so only read the first few pages.
   2. Extract the whole document.
   3. Extract the whole document, and all images in the document.

 PdfXmlParser - class to convert an XML document into Blocks and tokenizations.  This code was moved out of the Document class.

DocumentFactory - create a Document based on a list of Blocks, and tokenizations.

Document - delete a lot of code as a result of moving things around.  cleanLxml needs to be moved / rethought too.